requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.84.0
anyio==4.12.1
attrs==25.4.0
av==16.1.0
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
deprecation==2.1.0
distro==1.9.0
docstring_parser==0.17.0
eval_type_backport==0.3.1
fastapi==0.129.2
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.2.0
google-api-core==2.30.0
google-api-python-client==2.190.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
googleapis-common-protos==1.72.0
grpcio==1.78.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.1
jiter==0.13.0
livekit==1.1.2
livekit-agents==1.4.2
livekit-api==1.1.0
livekit-blingfire==1.1.0
livekit-plugins-cartesia==1.4.2
livekit-plugins-deepgram==1.4.2
livekit-plugins-elevenlabs==1.4.2
livekit-plugins-noise-cancellation==0.2.5
livekit-plugins-openai==1.4.2
livekit-plugins-sarvam==1.4.2
livekit-plugins-silero==1.4.2
livekit-protocol==1.1.2
markdown-it-py==4.0.0
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.1
nest-asyncio==1.6.0
numpy==2.4.2
onnxruntime==1.24.1
openai==2.21.0
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
packaging==26.0
pillow==12.1.1
prometheus_client==0.24.1
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.5
psutil==7.2.2
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyiceberg==0.11.0
PyJWT==2.11.0
pyparsing==3.3.2
pyroaring==1.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
pytz==2025.2
regex==2026.2.19
requests==2.32.5
rich==14.3.2
rsa==4.9.1
sarvamai==0.1.25
sentry-sdk==2.53.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.5
starlette==0.52.1
StrEnum==0.4.15
strictyaml==1.7.3
sympy==1.14.0
tenacity==9.1.4
tiktoken==0.12.0
tqdm==4.67.3
typer==0.24.0
types-protobuf==6.32.1.20251210
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.41.0
watchfiles==1.1.1
websockets==15.0.1
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0

asyncpg
pyairtable
python-dotenv
pydantic
requests
pandas
watchdog
dash
dash-bootstrap-components
plotly
python-jose
deepgram-sdk
pdfplumber
python-pptx
python-multipart
